Add files via upload
This is the first version. I might turn it into an actual app or have some default things or premade codes you can input shortcuts to. I just want to have a copy of this on the cloud.

Example of what you would input:
(first make the width and height of the screen something like 40x40)
0,1
1,2
2,0
2,1
2,2
done

This will make the classic Glider